requirements.txt
Flask==2.3.3
requests==2.31.0
PyPDF2==3.0.1
Pillow==10.4.0
pytesseract==0.3.10
openai==0.28.1
Werkzeug==2.3.7
plaid-python==9.1.0
python-dotenv==1.0.0 